Add files via upload
Business Case: In the recent past, the industry of wealth management has seen a lot of growth. Every individual or business actively searches for opportunities to get the maximum returns. However, in most of the cases, they either lack the skills to identify the right investment opportunity, or there is a shortage of time for finding these opportunities. Hence, this gave rise to the dedicated individuals who perform this task on behalf of the investors for a commission - Portfolio managers.
A portfolio manager makes investment decisions and carries out other related activities on behalf of vested investors. They work with a team of analysts and researchers, and their main objective is to realise the needs of the investor and suggest a suitable portfolio that meets all the expectations. They are responsible for establishing the best investment strategy, selecting appropriate investments along with the right allocation. However, in doing so, they face a lot of competition in the form of other portfolio managers and rival firms. Therefore, the portfolio manager has to use the available resources to provide the best solution to the investor.


Problem Statement:
Ms Alexandra Kolishnyick (aka Alexa) , daughter of a famous Ukrainian business tycoon, Intends to form an NGO along with her batchmates to support sub-Saharan African region. She has a huge savings from the money sent by parents for daily expense till date. She wants to invest the saved money in stocks which shows good returns and are also stable (Low risk).
Ms. Alexa wants a comprehensive report on all the metrics used to devise the investment strategy.